data/requirements.txt
PyMuPDF==1.22.5
Pillow==10.0.0
